css/css-fonts/parsing/font-stretch-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-stretch with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-fonts/#font-stretch-prop">
<meta name="assert" content="font-stretch supports the full grammar 'normal | ultra-condensed | extra-condensed | condensed | semi-condensed | semi-expanded | expanded | extra-expanded | ultra-expanded'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("font-stretch", "normal");
test_valid_value("font-stretch", "ultra-condensed");
test_valid_value("font-stretch", "extra-condensed");
test_valid_value("font-stretch", "condensed");
test_valid_value("font-stretch", "semi-condensed");
test_valid_value("font-stretch", "semi-expanded");
test_valid_value("font-stretch", "expanded");
test_valid_value("font-stretch", "extra-expanded");
test_valid_value("font-stretch", "ultra-expanded");
</script>
</body>
</html>
